refactored the message Listener to its own class instead of using an anonymous private class in the constructor and added a short comment to the new ClientListener class.
