+++ gamecore/src/main/java/com/mankomania/game/gamecore/client/NetworkClient.java
@@ -38,75 +38,7 @@ public class NetworkClient {
         this.messageHandler = new MessageHandler(this.client);
 
         client.start();
-        client.addListener(new Listener() {
-
-            @Override
-            public void received(Connection connection, Object object) {
-                Log.info(object.getClass().getSimpleName());
-
-                if (object instanceof PlayerConnected) {
-                    Log.info("player connected");
-                    Gdx.app.postRunnable(() -> ScreenManager.getInstance().switchScreen(Screen.LOBBY));
-
-                } else if (object instanceof ChatMessage) {
-                    ChatMessage response = (ChatMessage) object;
-                    //chat will be updated if message received
-                    ClientChat.addText(response.text);
-
-                    Log.info("[ChatMessage] received chat message (connection id: " + connection.getID() + "), text: '" + response.text + "'");
-                } else if (object instanceof InitPlayers) {
-                    // once game starts each player gets a list from server
-                    // and creates a hashMap with the IDs and player objects
-                    InitPlayers list = (InitPlayers) object;
-                    MankomaniaGame.getMankomaniaGame().getGameData().intPlayers(list.getPlayerIDs());
-
-                } else if (object instanceof Notification) {
-                    Notification notification = (Notification) object;
-                    MankomaniaGame.getMankomaniaGame().getNotifier().add(notification);
-
-                } else if (object instanceof StartGame) {
-                    /*
-                     * post a Runnable from networking thread to the libgdx rendering thread
-                     */
-                    Gdx.app.postRunnable(() -> ScreenManager.getInstance().switchScreen(Screen.MAIN_GAME));
-                    // once game starts each player gets a list from server
-                    // and creates a hashMap with the IDs and player objects
-                    StartGame gameStartedMessage = (StartGame) object;
-                    MankomaniaGame.getMankomaniaGame().getGameData().intPlayers(gameStartedMessage.getPlayerIds());
-                    MankomaniaGame.getMankomaniaGame().getGameData().setLocalPlayer(client.getID());
-
-                    Log.info("[GameStartedMessage] got GameStartedMessage, player array size: " + gameStartedMessage.getPlayerIds().size());
-                    Log.info("[GameStartedMessage] Initialized GameData with player id's");
-                } else if (object instanceof PlayerCanRollDiceMessage) {
-                    PlayerCanRollDiceMessage playerCanRollDiceMessage = (PlayerCanRollDiceMessage) object;
-
-                    Log.info("[PlayerCanRollDiceMessage] Player " + playerCanRollDiceMessage.getPlayerId() + " can roll the dice now!");
-
-                    messageHandler.gotPlayerCanRollDiceMessage(playerCanRollDiceMessage);
-                } else if (object instanceof MovePlayerToFieldMessage) {
-                    MovePlayerToFieldMessage movePlayerToFieldMessage = (MovePlayerToFieldMessage) object;
-
-                    Log.info("[MovePlayerToFieldMessage] Player " + movePlayerToFieldMessage.getPlayerId() + " got move to " + movePlayerToFieldMessage.getFieldToMoveTo() + " message");
-
-                    messageHandler.gotMoveToFieldMessage(movePlayerToFieldMessage);
-                } else if (object instanceof MovePlayerToIntersectionMessage) {
-                    MovePlayerToIntersectionMessage movePlayerToIntersectionMessage = (MovePlayerToIntersectionMessage) object;
-
-                    Log.info("[MovePlayerToIntersectionMessage] Player " + movePlayerToIntersectionMessage.getPlayerId() + " got to move to field " +
-                            movePlayerToIntersectionMessage.getFieldToMoveTo() + " and has to choose between path 1 = (" + movePlayerToIntersectionMessage.getSelectionOption1() +
-                            ") and path 2 = (" + movePlayerToIntersectionMessage.getSelectionOption2() + ")");
-
-                    messageHandler.gotMoveToIntersectionMessage(movePlayerToIntersectionMessage);
-                } else if (object instanceof MovePlayerToFieldAfterIntersectionMessage) {
-                    MovePlayerToFieldAfterIntersectionMessage movePlayerAfterIntersectionMsg = (MovePlayerToFieldAfterIntersectionMessage) object;
-
-                    Log.info("[MovePlayerToFieldAfterIntersectionMessage] Player " + movePlayerAfterIntersectionMsg.getPlayerId() + " got to move on the field " +
-                            movePlayerAfterIntersectionMsg.getFieldToMoveTo() + " directly after the intersection.");
-
-                    messageHandler.gotMoveAfterIntersectionMessage(movePlayerAfterIntersectionMsg);
-                }
-            }
-        });
+        client.addListener(new ClientListener(this.client, this.messageHandler)); // pass a new instance of the ClientListener to the client to handle all received messages.
     }
 
     public void tryConnectClient() {
